+++ pkg/auth/handler/webapp/settings_recovery_code.go
@@ -8,8 +8,10 @@ import (
 	"github.com/authgear/authgear-server/pkg/lib/config"
 	"github.com/authgear/authgear-server/pkg/lib/interaction/intents"
 	"github.com/authgear/authgear-server/pkg/lib/session"
+	"github.com/authgear/authgear-server/pkg/lib/web"
 	"github.com/authgear/authgear-server/pkg/util/httproute"
 	"github.com/authgear/authgear-server/pkg/util/httputil"
+	"github.com/authgear/authgear-server/pkg/util/secretcode"
 	"github.com/authgear/authgear-server/pkg/util/template"
 )
 
@@ -52,7 +54,7 @@ func (h *SettingsRecoveryCodeHandler) GetData(r *http.Request, rw http.ResponseW
 		for i, code := range codes {
 			recoveryCodes[i] = code.Code
 		}
-		viewModel.RecoveryCodes = formatRecoveryCodes(recoveryCodes)
+		viewModel.RecoveryCodes = secretcode.RecoveryCode.FormatCodes(recoveryCodes)
 	}
 
 	viewmodels.Embed(data, baseViewModel)
@@ -100,7 +102,7 @@ func (h *SettingsRecoveryCodeHandler) ServeHTTP(w http.ResponseWriter, r *http.R
 			return err
 		}
 
-		setRecoveryCodeAttachmentHeaders(w)
+		web.SetRecoveryCodeAttachmentHeaders(w)
 		h.Renderer.Render(w, r, TemplateWebDownloadRecoveryCodeTXT, data)
 		return nil
 	})
