+++ pkg/lib/web/plain_text.go
@@ -6,6 +6,11 @@ import (
 
 var TemplateWebRecoveryCodeTXT = template.RegisterPlainText("web/__recovery_code.txt")
 
+var TemplateWebDownloadRecoveryCodeTXT = template.RegisterPlainText(
+	"web/download_recovery_code.txt",
+	ComponentsPlainText...,
+)
+
 var ComponentsPlainText = []*template.PlainText{
 	TemplateWebRecoveryCodeTXT,
 }
