+++ pkg/lib/web/recovery_code.go
@@ -1,0 +1,14 @@
+package web
+
+import (
+	"mime"
+	"net/http"
+)
+
+func SetRecoveryCodeAttachmentHeaders(w http.ResponseWriter) {
+	// No need to use FormatMediaType because the value is constant.
+	w.Header().Set("Content-Type", "text/plain; charset=utf-8")
+	w.Header().Set("Content-Disposition", mime.FormatMediaType("attachment", map[string]string{
+		"filename": "recovery-codes.txt",
+	}))
+}
